Fixed Z80 sound driver compression
Closes #23

This is pretty hard to explain: AS outputs an object file, and it's
the s3p2bin tool that has to convert it to a ROM. In the object file,
the ROM is broken up into 'segments': for instance, the entire first
half of the Z80 sound driver is its own segment.

Each segment contains data, the data's length, and its address. This
address usually matches where the data should go in the ROM, but not
always: it actually specifies where in the relevant CPU's address
space it should go.

This becomes a problem when we have to deal with the Z80 sound
driver: the second half of the driver is stored in the cartridge at
0xF7760, but its segment address is 0x1300, because that's where it
gets loaded to in Z80 memory.

So, technically s3p2bin has no idea where the Z80 driver is meant to
go. Lucky for us, segments in the object file seem to be stored in
order, so it can just assume the Z80 driver goes after the last
segment, which *does* give us a proper address.

...But we can't rely on this either: the sound driver is split in
half, and between each half there is padding... but this padding
is completely ignored by s3p2bin, causing the second part of the
driver to be written to the wrong part of the cartridge. This is
why the sound driver is broken in recent builds.

The reason this padding is ignored is because it basically doesn't
exist: the padding is created by using an '!org' directive, but the
way it generates padding is merely by changing the address value of
the next segment... which gets overwritten since the next segment
is a Z80 segment.

Previously, this disassembly seemed to work by sheer accident: these
'!org's used to be followed by 'align0' macros, which also created
padding, except by creating their own segment instead of merely
meddling with the one after it. This allowed s3p2bin to properly
insert the second Z80 segment at the right ROM address.

This isn't an ideal solution, however: the 'align0' macros were meant
to pad to the nearest 0x10-byte boundary, so in theory if the align0
macro just happened to appear on an already-aligned address, it would
never generate a new segment, causing the ROM to break at seemingly
random.

To fix all of this for good, I've changed how padding works:
now, before the Z80 code even begins, an 'org' macro is used (unlike
'!org', 'org' works by creating a new segment like 'align0' does),
allocating all of the room needed for both the Z80 code and its
padding. Then, when s3p2bin detects a Z80 segment, it goes back and
overwrites the previous segment with it, instead of tacking it onto
the end.

Right now, the Windows s3p2bin is the only one with the new
behaviour. I'll have to recompile the Linux one in another commit.