Migrate the build system to Lua.
This is the same build system that's used by the Sonic 1 and Sonic 2
disassemblies. This allows Windows and non-Windows OSes to use the
exact same build scripts, avoiding duplication (previously there were
Batch scripts for Windows users and Python scripts for everybody
else). A Lua runtime executable is provided for Windows users so that
they don't need to install anything in order to build the
disassembly.

Additionally, bringing the build system up to date with Sonic 1 and 2
provides partial support for 32-bit x86 Linux and 64-bit x86 FreeBSD
(s3p2bin binaries are needed for full support).

The Travis script was purged in this commit. Hopefully it wasn't
still being used.