Disassemble Sonic 3's sound driver.
This has been sat incomplete in my local repo for over a year. Issues
relating to S3's driver being uncompressed while S&K's driver was
compressed were blocking this, but ever since the new p2bin has been
added, this is no longer a problem. So, here it is.